break dependency cycle AsgResource <-> AutoScalingGroup
